Added context parameters and nested type aliases support
Also, removed context receivers support, because it was replaced by context receivers parameters

#KTL-2182
