+++ common/src/main/kotlin/component/KotlinEnvironment.kt
@@ -50,7 +50,8 @@ class KotlinEnvironment(
       "-opt-in=kotlin.experimental.ExperimentalTypeInference",
       "-opt-in=kotlin.uuid.ExperimentalUuidApi",
       "-opt-in=kotlin.io.encoding.ExperimentalEncodingApi",
-      "-Xcontext-receivers",
+      "-Xcontext-parameters",
+      "-Xnested-type-aliases",
       "-Xreport-all-warnings",
       "-Wextra",
       "-XXLanguage:+ExplicitBackingFields",
