+++ src/test/kotlin/com/compiler/server/KotlinFeatureSince160.kt
@@ -71,35 +71,4 @@ class KotlinFeatureSince160 : BaseExecutorTest() {
       contains = "<outStream>[1, 2, 3, 5]\n</outStream>"
     )
   }
-
-  @Test
-  fun `Prototype of context receivers`() {
-    run(
-      code = """
-        class Logger {
-          fun info(message: String) = println(message)
-        }
-
-        interface LoggingContext {
-          val log: Logger
-        }
-
-        context(LoggingContext)
-        fun executeTask() {
-          log.info("Complete")
-        }
-
-        fun main() {
-          val loggingContext = object: LoggingContext {
-            override val log = Logger()
-          }
-            
-          with(loggingContext) {
-            executeTask()
-          }
-        }
-      """.trimIndent(),
-      contains = "Complete"
-    )
-  }
 }
